gestion des groupes (ajout, modif, suppression,avec cascade)

+++ Annuaire/src/annuaire/web/GroupController.java
@@ -1,6 +1,7 @@
 package annuaire.web;
 
 import java.util.Collection;
+import java.util.HashSet;
 
 import javax.validation.Valid;
 
@@ -13,6 +14,7 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 
 import annuaire.model.Group;
+import annuaire.model.Person;
 import annuaire.services.IDAOGroup;
 
 @Controller()
@@ -21,10 +23,10 @@ public class GroupController { //TODO gerer interceptor
 
 	@Autowired()
 	User user;
-	
+
 	@Autowired
 	IDAOGroup daoGroup;
-	
+
 	@RequestMapping(value = "/groups.htm")
 	public String list() {
 		return "listGroup";
@@ -34,9 +36,6 @@ public class GroupController { //TODO gerer interceptor
 	public Collection<Group> groups() {
 		return daoGroup.findAllGroups();
 	}
-
-
-
 
 	@ModelAttribute("group")
 	public Group newGroup(
@@ -48,6 +47,7 @@ public class GroupController { //TODO gerer interceptor
 
 		Group g = new Group();
 		g.setGroupname("");
+		g.setPersons(new HashSet<Person>());
 		return g;
 	}
 
@@ -58,11 +58,11 @@ public class GroupController { //TODO gerer interceptor
 			return "redirect:groups.htm";
 		}
 
-		if(!user.getGroups().contains("ADMINISTRATEUR")) {
-			return "redirect:groups.htm";
+		if(user.isAdmin()) {
+			return "groupForm";
 		}
+		return "redirect:groups.htm";
 
-		return "groupForm";
 	}
 
 	@RequestMapping(value = "/edition.htm", method = RequestMethod.POST)
@@ -70,23 +70,24 @@ public class GroupController { //TODO gerer interceptor
 		if(result.hasErrors()) {
 			return "groupForm";
 		}
-		
+
 		if(g == null || g.getGroupname().isEmpty()) {
 			return "redirect:groups.htm";
 		}
 
-		if(!user.getGroups().contains("ADMINISTRATEUR")) {
-			return "redirect:groups.htm";
+		if(user.isAdmin()) {
+			daoGroup.saveGroup(g);
 		}
-
-		daoGroup.saveGroup(g);
-
 		return "redirect:groups.htm";
+
 	}
 
-	@RequestMapping(value = "/add.htm")
+	@RequestMapping(value = "/add.htm", method = RequestMethod.GET)
 	public String addGroup() {
-		return "groupAddition";
+		if (user.isAdmin()){
+			return "groupAddition";
+		}
+		return "redirect:groups.htm";
 	}
 
 	@RequestMapping(value = "/add.htm", method = RequestMethod.POST)
@@ -96,12 +97,9 @@ public class GroupController { //TODO gerer interceptor
 			return "groupAddition";
 		}
 
-		if(!user.getGroups().contains("ADMINISTRATEUR")) {
-			return "redirect:groups.htm";
+		if (user.isAdmin()){
+			daoGroup.addGroup(g);
 		}
-		
-		daoGroup.addGroup(g);
-
 		return "redirect:groups.htm";
 	}
 
@@ -112,27 +110,15 @@ public class GroupController { //TODO gerer interceptor
 			return "redirect:groups.htm";
 		}
 
-		daoGroup.deleteGroup(g.getGroupname());
-
+		if (user.isAdmin()){
+			daoGroup.deleteGroup(g.getGroupname());
+		}
 		return "redirect:groups.htm";
 	}
-	
+
 	@ModelAttribute("user")
 	public User newUser() {
 		return user;
 	}
-	
-//	@ModelAttribute("groupsList")
-//	public Map<String, String> productTypes() {
-//		
-//		ArrayList<Group> groups = (ArrayList<Group>) daoGroup.findAllGroups();
-//		
-//	    Map<String, String> types = new LinkedHashMap<>();
-//	    
-//	    for(Group g: groups){
-//	    	types.put(g.getGroupname(), g.getGroupname());
-//	    }
-//	    return types;
-//	}
-	
+
 }
